Update Credentials functions arg order to match Signatures arg order
Before this, functions on the Credentials module had the following
argument order:

    Credentials.generate_credential_scope_v4(service, datetime, config)

and functions on the Signatures module had the following argument order

    Signatures.generate_signature_v4(service, config, datetime, string_to_sign)

When calling these functions close to each other, it looked weird having 
these arguments in different orders. This change updates the argument
order on the functions in the Credentials module to match the signatures
module.

    Credentials.generate_credential_scope_v4(service, config, datetime)