Turn empty config maps to nil to resolve correctly
Enum.find(&(&1)) is used to find the first truthy
(non nil) configuration when resolving multiple sources.

retrieve_runtime_value correctly returns nil for :system
but returns empty maps for :awscli and :instance_role, which
are truthy.

This means that with the default configuration with no
config/credentials file available for awscli the configuration
will be set to an empty map instead of falling back to
:instance_role.