fixed CustomUserData if it’s nil
I have run into a strange and confusing situation with SNS and ex_aws. Specifically when registering a device with `SNS.create_platform_endpoint`. Before this PR, a nil value for `custom_user_data` gets converted to an empty string `””` on AWS side.

Additionally, a device registered with another client (aws cli for example) that truly does have no custom user data incorrectly shows up with the custom_user_data as `””` with ex_aws.

This causes a problem because when re-registering a device (which you are allowed to do), and the custom user data does not match you will get an AWS error.

This PR prevents a nil `custom_user_data` param from getting sent to AWS (which causes the empty string to be set). It also removes the custom_user_data param when parsing a `GetEndpointAttributesResponse` response if there is no custom_user_data (instead of setting it to an empty string).