Pass Headers to Request
This change allows HTTP headers to be passed in. This was originally questioned when it was merged in here: https://github.com/ex-aws/ex_aws/pull/286#discussion_r90781530 but wasn't necessary for the use case at the time. However, attempting to use this with Cloudfront (and presumably any other service that utilizes ETag) makes this necessary.