Support passing headers
Passing Headers, such as `If-Match`, was previously causing issues. This is because the following code was executing when passing an array of header values:

```
  def retrieve_runtime_value(values, config) when is_list(values) do
    values
    |> Stream.map(&retrieve_runtime_value(&1, config))
    |> Enum.find(& &1)
  end
```

which destructured the data from `[{ "header-name", "header-value"}]` to `{ "header-name", "header-value"}` (note that it is no longer a list containing a tuple, but rather simply a tuple.

This caused issues later when trying to build up the headers to send in the request. Abbreviated for simplicity, we had:

```elixir
headers = { "some-key", "some-value" }
headers = [ defaultHeaders | headers ]
```

Because `headers` was no longer a `List`, the `headers` variable now contains an improper List.

This change passes `config.headers` through to the underlying functions without destructuring it first.