Fix Req example HttpClient implementation for Req 0.4+
This builds on work in #1015 adding a bit of code to adjust the header
format to the expected list of 2-element tuples.

Req 0.4+ changed the header format to be a map where the key is a string
and the value is a list of strings, e.g.:

```elixir
%{
  "content-length" => ["12345"]
}
```

This change makes the current example fail whenever you try to perform a
streaming operation because this code in `ExAws.S3.Download` assumes a
single value:

```elixir
defp get_file_size(bucket, path, config) do
  %{headers: headers} = ExAws.S3.head_object(bucket, path) |> ExAws.request!(config)

  headers
  |> Enum.find(fn {k, _} -> String.downcase(k) == "content-length" end)
  |> elem(1)
  |> String.to_integer()
end
```

With the proposed change, the example implementation should handle both new and old
Req versions; though I've only tested with a recent one.