Don't set content-length header for HEAD requests with empty body
This extra function clause prevents the `content-length` (equals 0)
header from being added to the headers of a `HEAD` request with an empty
body. This is compliant with [Amazon's specification of the HeadObject
API endpoint] that doesn't specify the `content-length` header as part
of the URI Request Parameters. As a benefit, this makes ExAws and
ExAws.S3 work with [Garage](https://garagehq.deuxfleurs.fr/), which,
unlike [Minio](https://min.io), doesn't gracefully handle the extraneous
presence of the `{"content-length", "0"}` header in the request headers
of the HeadObject call.