Add Bedrock Agent endpoints
Map the Bedrock Agent and Bedrock Agent Runtime endpoints. This allows
to start using the services with the `ex_aws` library.

The PR is identical to the one which introduced endpoints for Bedrock
and Bedrock Runtime: https://github.com/ex-aws/ex_aws/pull/1023.

The list of supported endpoints relies on
https://docs.aws.amazon.com/general/latest/gr/bedrock.html#bedrock_region.